% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemddenoise.R
\name{eemddenoise}
\alias{eemddenoise}
\title{Denosing using EEMD method}
\usage{
eemddenoise(
  xt,
  tt = NULL,
  cv.index,
  cv.level,
  cv.tol = 0.1^3,
  cv.maxiter = 20,
  by.imf = FALSE,
  trials.dir = "EEMD",
  noise.amp = sd(xt) * 0.1,
  trials = 5,
  nimf = 5,
  tol = sd(xt) * 0.1^2,
  max.sift = 20,
  stop.rule = "type1",
  boundary = "periodic",
  max.imf = 10
)
}
\arguments{
\item{xt}{The input signal.}
}
\value{
The smoothed signal.
}
\description{
This function smooths the input signal using the EEMD method.
}
