% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRS_helpers.R
\name{findPeaks}
\alias{findPeaks}
\title{search peaks for each spectrum}
\usage{
findPeaks(spc, size = 9, level = 0.1)
}
\arguments{
\item{spc}{an intensities vector}

\item{size}{need to be odd number}

\item{level}{adjustable constrain}
}
\description{
search peaks for each spectrum
}
