% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rammanome-class.R
\name{get.nearest.dataset}
\alias{get.nearest.dataset}
\title{Get the nearest dataset from a Ramanome object}
\usage{
get.nearest.dataset(object)
}
\arguments{
\item{object}{A Ramanome object.}
}
\value{
The nearest dataset.
}
\description{
This function retrieves the nearest dataset from a Ramanome object.
}
