% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvd.R
\name{hvd}
\alias{hvd}
\title{Hilbert Vibration Decomposition}
\usage{
hvd(x, n, fp)
}
\arguments{
\item{x}{The input signal.}

\item{n}{The number of components to decompose.}

\item{fp}{Lowpass filter cutoff frequency.}
}
\value{
A list containing the decomposed components, amplitudes,
angular frequencies, and relative standard deviations.
}
\description{
This function implements a variant of the Hilbert Vibration Decomposition.
}
