% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvd.R
\name{ilpf}
\alias{ilpf}
\title{Ideal Lowpass Filter}
\usage{
ilpf(x, fp)
}
\arguments{
\item{x}{The input signal.}

\item{fp}{Filter cutoff frequency (0 < fp < 0.5).}
}
\value{
The filtered signal.
}
\description{
This function applies an ideal lowpass filter to the input signal.
}
