% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvd.R
\name{inst}
\alias{inst}
\title{Instantaneous Properties}
\usage{
inst(x, Fs)
}
\arguments{
\item{x}{The input signal.}

\item{Fs}{Sampling frequency.}
}
\value{
A list containing the amplitude, frequency, and phase.
}
\description{
This function calculates the instantaneous amplitude, frequency,
and phase of a signal.
}
