% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRS_helpers.R
\name{removeCosmic}
\alias{removeCosmic}
\title{remove cosmic rays for each spectrum}
\usage{
removeCosmic(spc)
}
\arguments{
\item{spc}{an intensities vector}
}
\description{
remove cosmic rays for each spectrum
}
