% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRS_preprocess.R
\name{SCRS_preprocess}
\alias{SCRS_preprocess}
\title{Preprocess SCRS data}
\usage{
SCRS_preprocess(input_dir, output, meta_fp, txt_filename)
}
\arguments{
\item{input_dir}{A directory}

\item{output}{A directory}

\item{meta_fp}{A TSV file}

\item{txt_filename}{A prefix of file names}
}
\value{
Results were put into output directory
}
\description{
Baseline and normalization for SCRS data
}
