% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvd.R
\name{synchdem}
\alias{synchdem}
\title{Synchronous Demodulation}
\usage{
synchdem(x, omega_i, fp)
}
\arguments{
\item{x}{The input signal.}

\item{omega_i}{The frequency vector.}

\item{fp}{Lowpass filter cutoff frequency.}
}
\value{
A list containing the demodulated signal, amplitude, and phase.
}
\description{
This function performs synchronous demodulation of a signal.
}
